feat: changed stats and tps calculations; added debug command and better logging

+++ src/main/java/com/dragonminez/client/gui/MastersSkillsScreen.java
@@ -221,7 +221,7 @@ public class MastersSkillsScreen extends BaseMenuScreen {
 
 		if (!statsData.getSkills().hasSkill(selectedSkill) || skill.getLevel() == 0) {
 			int cost = getUpgradeCost(selectedSkill, 0);
-			int currentTPS = statsData.getResources().getTrainingPoints();
+			float currentTPS = statsData.getResources().getTrainingPoints();
 			boolean canAfford = currentTPS >= cost;
 			if (cost == -1 || cost == Integer.MAX_VALUE) return;
